Prep for upgrade to 1.2.0

+++ src/ft_greater.c
@@ -3,14 +3,14 @@
 /*                                                        :::      ::::::::   */
 /*   ft_greater.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: iwordes <marvin@42.fr>                     +#+  +:+       +#+        */
+/*   By: iwordes <iwordes@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/10/02 16:42:14 by iwordes           #+#    #+#             */
-/*   Updated: 2016/10/05 10:02:16 by iwordes          ###   ########.fr       */
+/*   Updated: 2016/12/14 09:13:27 by iwordes          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-long long	ft_greater(long long num1, long long num2)
+intmax_t	ft_greater(intmax_t num1, intmax_t num2)
 {
 	return ((num1 > num2) ? num1 : num2);
 }
