+++ src/ft_itoa.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   ft_itoa.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: iwordes <marvin@42.fr>                     +#+  +:+       +#+        */
+/*   By: iwordes <iwordes@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/10/02 16:42:46 by iwordes           #+#    #+#             */
-/*   Updated: 2016/10/05 09:56:23 by iwordes          ###   ########.fr       */
+/*   Updated: 2016/12/14 09:15:55 by iwordes          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char	*ft_itoa(int integer)
+char	*ft_itoa(intmax_t integer)
 {
 	return (ft_itoa_base(integer, 10));
 }
