+++ src/ft_lesser.c
@@ -3,14 +3,14 @@
 /*                                                        :::      ::::::::   */
 /*   ft_lesser.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: iwordes <marvin@42.fr>                     +#+  +:+       +#+        */
+/*   By: iwordes <iwordes@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/10/02 16:42:51 by iwordes           #+#    #+#             */
-/*   Updated: 2016/10/05 10:04:23 by iwordes          ###   ########.fr       */
+/*   Updated: 2016/12/14 09:16:14 by iwordes          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-long long		ft_lesser(long long num1, long long num2)
+intmax_t	ft_lesser(intmax_t num1, intmax_t num2)
 {
 	return ((num1 < num2) ? num1 : num2);
 }
